Hacky way to make merge function compile
Simply addressed the two compile errors, that we only passed in one
argument to the Constant merge function constructor, and that calling a
non-constant function for a constant parameter is not allowed. We need
to determine if the two changes (changign the constructor to take one
argument and marking the operator() function as const, make conceptual
sense.